Just splitting the one file into topic files
